requirements.txt
# Core Scientific Libraries
numpy>=1.21.0,<2.0.0
pandas>=1.3.0
scipy>=1.7.0
scikit-learn>=1.0.0

# Network Packet Capture
scapy>=2.4.5

# HTTP Requests
requests>=2.28.0

# Web Dashboard
Flask>=2.3.0
Flask-SocketIO>=5.3.0
Flask-CORS>=4.0.0
python-socketio>=5.9.0
eventlet>=0.33.0

# Database
# SQLite is included with Python

# Utilities
python-dotenv>=1.0.0

# === UPGRADE PHASE 1: High-Speed Capture ===
# PyShark (optional - for high-speed capture)
# pyshark>=0.6.0

# LibPCAP (optional - for ultra-fast capture)
# pcapy>=0.11.5
# impacket>=0.11.0

# === UPGRADE PHASE 2: Distributed Processing ===
# Kafka
kafka-python>=2.0.2

# Redis
redis>=5.0.0

# === UPGRADE PHASE 3: Deep Learning ===
# TensorFlow
tensorflow>=2.13.0

# GPU Acceleration (optional - requires CUDA)
# cupy-cuda11x>=12.0.0
# cudf-cuda11x>=23.0.0
# cuml-cuda11x>=23.0.0

# === UPGRADE PHASE 4: Monitoring & Metrics ===
# Prometheus
prometheus-client>=0.19.0

# === Additional Dependencies ===
# Job scheduling
joblib>=1.3.0

